Safari-611.1.10/LayoutTests/fast/forms/placeholder-stripped-expected.txt

CR and LF in placeholder values should be stripped.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.visiblePlaceholder(input0) is "first line second line"
PASS internals.visiblePlaceholder(input1) is ""
PASS internals.visiblePlaceholder(textarea0) is "first line second line"
PASS internals.visiblePlaceholder(textarea1) is ""
PASS successfullyParsed is true

TEST COMPLETE
